+++ tests/testthat/test-LoadCellGraphs.R
@@ -65,9 +65,10 @@ for (assay_version in c("v3", "v5")) {
     layouts <- seur_obj_mpx_precomputed[["mpxCells"]]@cellgraphs[[1]]@layout
     expect_equal(dim(layouts[[1]]), c(2470, 3))
     expect_equal(
-      rownames(layouts[[1]]),
+      Cells(seur_obj_mpx_precomputed[["mpxCells"]]@cellgraphs[[1]]),
       seur_obj_mpx_precomputed[["mpxCells"]]@cellgraphs[[1]]@cellgraph %>% pull(name)
     )
+    expect_lt(.row_names_info(layouts[[1]]), 0L)
     expect_equal(
       layouts[[1]] %>% head() %>% tibble::as_tibble(),
       structure(
@@ -104,9 +105,10 @@ for (assay_version in c("v3", "v5")) {
     layouts <- seur_obj_pna[["PNA"]]@cellgraphs[[1]]@layout
     expect_equal(dim(layouts[[1]]), c(43543, 3))
     expect_equal(
-      rownames(layouts[[1]]),
+      Cells(seur_obj_pna[["PNA"]]@cellgraphs[[1]]),
       seur_obj_pna[["PNA"]]@cellgraphs[[1]]@cellgraph %>% pull(name)
     )
+    expect_lt(.row_names_info(layouts[[1]]), 0L)
     expect_equal(
       layouts[[1]] %>% head() %>% tibble::as_tibble(),
       structure(